requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.21
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing-extensions==4.13.2
